refactor: centralize clean_for_pdf helper
