+++ tests/test_clean_for_pdf.py
@@ -1,4 +1,4 @@
-from src.assignment_ui import clean_for_pdf
+from src.pdf_utils import clean_for_pdf
 
 
 def test_clean_for_pdf_preserves_unicode_and_strips_control_chars():
